LayoutTests/fast/notifications/notifications-document-close-crash-expected.txt

CONSOLE MESSAGE: line 36: Uncaught InvalidStateError (DOM Exception 11): The object is in an invalid state.
PASS
